Moed ArrayStaet subclasses into ArrayState.h and ArrayState.cpp.
