+++ src/vsg/traversals/ArrayState.cpp
@@ -13,11 +13,14 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/commands/BindIndexBuffer.h>
 #include <vsg/commands/BindVertexBuffers.h>
 #include <vsg/io/Options.h>
+#include <vsg/maths/sample.h>
 #include <vsg/nodes/Geometry.h>
 #include <vsg/nodes/VertexIndexDraw.h>
 #include <vsg/state/GraphicsPipeline.h>
 #include <vsg/state/InputAssemblyState.h>
 #include <vsg/state/VertexInputState.h>
+#include <vsg/state/DescriptorSet.h>
+#include <vsg/state/DescriptorImage.h>
 #include <vsg/traversals/ArrayState.h>
 
 using namespace vsg;
@@ -176,4 +179,241 @@ void NullArrayState::apply(const vsg::vec3Array&)
 void NullArrayState::apply(const vsg::Data&)
 {
     vertices = {};
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// PositionArrayState
+//
+PositionArrayState::PositionArrayState()
+{
+}
+
+PositionArrayState::PositionArrayState(const PositionArrayState& rhs) :
+    Inherit(rhs),
+    position_attribute_location(rhs.position_attribute_location),
+    positionAttribute(rhs.positionAttribute)
+{
+}
+
+PositionArrayState::PositionArrayState(const ArrayState& rhs) :
+    Inherit(rhs)
+{
+}
+
+ref_ptr<ArrayState> PositionArrayState::clone()
+{
+    return PositionArrayState::create(*this);
+}
+
+ref_ptr<ArrayState> PositionArrayState::clone(ref_ptr<ArrayState> arrayState)
+{
+    return PositionArrayState::create(*arrayState);
+}
+
+void PositionArrayState::apply(const VertexInputState& vas)
+{
+    getAttributeDetails(vas, vertex_attribute_location, vertexAttribute);
+    getAttributeDetails(vas, position_attribute_location, positionAttribute);
+}
+
+ref_ptr<const vec3Array> PositionArrayState::vertexArray(uint32_t instanceIndex)
+{
+    auto positions = arrays[positionAttribute.binding].cast<vec3Array>();
+
+    if (positions && (instanceIndex < positions->size()))
+    {
+        auto position = positions->at(instanceIndex);
+        auto new_vertices = vsg::vec3Array::create(vertices->size());
+        auto src_vertex_itr = vertices->begin();
+        for(auto& v : *new_vertices)
+        {
+            v = *(src_vertex_itr++) + position;
+        }
+        return new_vertices;
+    }
+
+    return vertices;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// DisplacementMapArrayState
+//
+DisplacementMapArrayState::DisplacementMapArrayState()
+{
+}
+
+DisplacementMapArrayState::DisplacementMapArrayState(const DisplacementMapArrayState& rhs) :
+    Inherit(rhs)
+{
+}
+
+DisplacementMapArrayState::DisplacementMapArrayState(const ArrayState& rhs) :
+    Inherit(rhs)
+{
+}
+
+ref_ptr<ArrayState> DisplacementMapArrayState::clone()
+{
+    return DisplacementMapArrayState::create(*this);
+}
+
+ref_ptr<ArrayState> DisplacementMapArrayState::clone(ref_ptr<ArrayState> arrayState)
+{
+    return DisplacementMapArrayState::create(*arrayState);
+}
+
+void DisplacementMapArrayState::apply(const DescriptorImage& di)
+{
+    if (di.imageInfoList.size() >= 1)
+    {
+        auto& imageInfo = *di.imageInfoList[0];
+        if (imageInfo.imageView && imageInfo.imageView->image)
+        {
+            displacementMap = imageInfo.imageView->image->data.cast<floatArray2D>();
+            sampler = imageInfo.sampler;
+        }
+    }
+}
+
+void DisplacementMapArrayState::apply(const DescriptorSet& ds)
+{
+    for(auto& descriptor : ds.descriptors)
+    {
+        if (descriptor->dstBinding == dm_binding)
+        {
+            descriptor->accept(*this);
+            break;
+        }
+    }
+}
+
+void DisplacementMapArrayState::apply(const BindDescriptorSet& bds)
+{
+    if (bds.firstSet == dm_set)
+    {
+        apply(*bds.descriptorSet);
+    }
+}
+
+void DisplacementMapArrayState::apply(const BindDescriptorSets& bds)
+{
+    if (bds.firstSet <= dm_set && dm_set < (bds.firstSet+ + static_cast<uint32_t>(bds.descriptorSets.size())))
+    {
+        apply(*bds.descriptorSets[dm_set - bds.firstSet]);
+    }
+}
+
+void DisplacementMapArrayState::apply(const VertexInputState& vas)
+{
+    getAttributeDetails(vas, vertex_attribute_location, vertexAttribute);
+    getAttributeDetails(vas, normal_attribute_location, normalAttribute);
+    getAttributeDetails(vas, texcoord_attribute_location, texcoordAttribute);
+}
+
+ref_ptr<const vec3Array> DisplacementMapArrayState::vertexArray(uint32_t /*instanceIndex*/)
+{
+    if (displacementMap)
+    {
+        auto normals = arrays[normalAttribute.binding].cast<vec3Array>();
+        auto texcoords = arrays[texcoordAttribute.binding].cast<vec2Array>();
+        if (texcoords->size() != vertices->size()) return {};
+        if (normals->size() != vertices->size()) return {};
+
+        auto new_vertices = vsg::vec3Array::create(vertices->size());
+        auto src_vertex_itr = vertices->begin();
+        auto src_teccoord_itr = texcoords->begin();
+        auto src_normal_itr = normals->begin();
+        vec2 tc_scale(static_cast<float>(displacementMap->width())-1.0f, static_cast<float>(displacementMap->height())-1.0f);
+
+        // if no sampler is assigned fallback to use default constructed Sampler
+        if (!sampler) sampler = Sampler::create();
+
+        for(auto& v : *new_vertices)
+        {
+            auto& tc = *(src_teccoord_itr++);
+            auto& n = *(src_normal_itr++);
+            float d = sample(*sampler, *displacementMap, tc);
+            v = *(src_vertex_itr++) + n * d;
+        }
+        return new_vertices;
+    }
+
+    return vertices;
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// PositionAndDisplacementMapArrayState
+//
+PositionAndDisplacementMapArrayState::PositionAndDisplacementMapArrayState()
+{
+}
+
+PositionAndDisplacementMapArrayState::PositionAndDisplacementMapArrayState(const PositionAndDisplacementMapArrayState& rhs) :
+    Inherit(rhs)
+{
+}
+
+PositionAndDisplacementMapArrayState::PositionAndDisplacementMapArrayState(const ArrayState& rhs) :
+    Inherit(rhs)
+{
+}
+
+ref_ptr<ArrayState> PositionAndDisplacementMapArrayState::clone()
+{
+    return PositionAndDisplacementMapArrayState::create(*this);
+}
+
+ref_ptr<ArrayState> PositionAndDisplacementMapArrayState::clone(ref_ptr<ArrayState> arrayState)
+{
+    return PositionAndDisplacementMapArrayState::create(*arrayState);
+}
+
+void PositionAndDisplacementMapArrayState::apply(const VertexInputState& vas)
+{
+    getAttributeDetails(vas, vertex_attribute_location, vertexAttribute);
+    getAttributeDetails(vas, normal_attribute_location, normalAttribute);
+    getAttributeDetails(vas, texcoord_attribute_location, texcoordAttribute);
+    getAttributeDetails(vas, position_attribute_location, positionAttribute);
+}
+
+ref_ptr<const vec3Array> PositionAndDisplacementMapArrayState::vertexArray(uint32_t instanceIndex)
+{
+    auto positions = arrays[positionAttribute.binding].cast<vec3Array>();
+
+    vec3 position;
+    if (positions && (instanceIndex < positions->size()))
+    {
+        position = positions->at(instanceIndex);
+    }
+
+    if (displacementMap)
+    {
+        auto normals = arrays[normalAttribute.binding].cast<vec3Array>();
+        auto texcoords = arrays[texcoordAttribute.binding].cast<vec2Array>();
+        if (texcoords->size() != vertices->size()) return {};
+        if (normals->size() != vertices->size()) return {};
+
+        auto new_vertices = vsg::vec3Array::create(vertices->size());
+        auto src_vertex_itr = vertices->begin();
+        auto src_teccoord_itr = texcoords->begin();
+        auto src_normal_itr = normals->begin();
+        vec2 tc_scale(static_cast<float>(displacementMap->width())-1.0f, static_cast<float>(displacementMap->height())-1.0f);
+
+        // if no sampler is assigned fallback to use default constructed Sampler
+        if (!sampler) sampler = Sampler::create();
+
+        for(auto& v : *new_vertices)
+        {
+            auto& tc = *(src_teccoord_itr++);
+            auto& n = *(src_normal_itr++);
+            float d = sample(*sampler, *displacementMap, tc);
+            v = *(src_vertex_itr++) + n * d + position;
+        }
+        return new_vertices;
+    }
+
+    return vertices;
 }
